REST: Prevent delete_item from deleting normal revision type

+++ revisions-extended/includes/rest-revisions-controller.php
@@ -218,7 +218,22 @@ class REST_Revisions_Controller extends WP_REST_Revisions_Controller {
 	 */
 	public function delete_item( $request ) {
 		add_filter( 'rest_prepare_revision', array( $this, 'filter_rest_prepare_revision' ), 10, 3 );
+
+		$revision = $this->get_revision( $request['id'] );
+		if ( is_wp_error( $revision ) ) {
+			return $revision;
+		}
+
+		if ( ! in_array( $revision->post_status, wp_list_pluck( get_revision_statuses(), 'name' ), true ) ) {
+			return new WP_Error(
+				'rest_invalid_revision',
+				__( 'This type of revision cannot be deleted with this endpoint. Use the Core endpoint instead.', 'revisions-extended' ),
+				array( 'status' => 403 )
+			);
+		}
+
 		$response = parent::delete_item( $request );
+
 		remove_filter( 'rest_prepare_revision', array( $this, 'filter_rest_prepare_revision' ) );
 
 		return $response;
